clean:promisify the procees of retrieving lines
