+++ 04.class/main.js
@@ -1,7 +1,7 @@
 #! /usr/bin/env node
 
 import minimist from "minimist";
-import readline from "readline";
+import promisifiedReadlineFunctions from "./promisified-readline-functions.js";
 import Manager from "./manager.js";
 
 const options = minimist(process.argv.slice(2), {
@@ -15,25 +15,13 @@ const options = minimist(process.argv.slice(2), {
 const memoManager = new Manager();
 await memoManager.createTable();
 
-if (!options.list && !options.reference && !options.delete) {
-  const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout,
-  });
-  const lines = [];
-  rl.on("line", (line) => {
-    lines.push(line);
-  });
-  rl.on("SIGINT", () => {
-    process.exit(130);
-  });
-  rl.on("close", async () => {
-    await memoManager.add(lines[0], lines.slice(1).join("\n"));
-  });
-} else if (options.list) {
+if (options.list) {
   memoManager.showList();
 } else if (options.reference) {
   memoManager.refer();
 } else if (options.delete) {
   memoManager.delete();
+} else {
+  const lines = await promisifiedReadlineFunctions.inputLines();
+  await memoManager.add(lines[0], lines.slice(1).join("\n"));
 }
